+++ tests/Fixtures/app/AppKernel.php
@@ -1,14 +1,13 @@
-<?php
+<?php /** @noinspection PhpIllegalPsrClassPathInspection */
 
 declare(strict_types=1);
 
 use ApiPlatform\Symfony\Bundle\ApiPlatformBundle;
 use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
-use Doctrine\Bundle\FixturesBundle\DoctrineFixturesBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
 use Symfony\Bundle\MonologBundle\MonologBundle;
-use Symfony\Bundle\SecurityBundle\SecurityBundle;
+use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpFoundation\Session\SessionFactory;
@@ -24,7 +23,9 @@ class AppKernel extends Kernel
             new FrameworkBundle(),
             new DoctrineBundle(),
             new MonologBundle(),
+            new TwigBundle(),
             new ApiPlatformBundle(),
+            new Vrok\SymfonyAddons\VrokSymfonyAddonsBundle(),
         ];
 
         return $bundles;
@@ -35,13 +36,13 @@ class AppKernel extends Kernel
         return __DIR__;
     }
 
-    protected function configureContainer(ContainerBuilder $c, LoaderInterface $loader): void
+    protected function configureContainer(ContainerBuilder $container, LoaderInterface $loader): void
     {
-        $c->setParameter('kernel.project_dir', __DIR__);
+        $container->setParameter('kernel.project_dir', __DIR__);
 
         $loader->load(__DIR__.'/config/config.yaml');
 
-        $c->prependExtensionConfig('framework', [
+        $container->prependExtensionConfig('framework', [
             'property_access'      => ['enabled' => true],
             'secret'               => 'symfony.vrok',
             'validation'           => ['enable_annotations' => true],